pyproject.toml
[project]
name = "sig-consulting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "nbformat>=5.10.4",
    "pandas>=2.3.3",
    "gspread>=6.1.2",
    "google-auth>=2.34.0",
    "python-dotenv>=1.1.1",
    "networkx>=3.5",
    "matplotlib>=3.10.6",
]
